requirements.txt
fastapi
uvicorn
openai
python-dotenv
langchain
langgraph
pydantic
moviepy
gtts
diffusers
transformers
torch
pillow
faiss-cpu
psycopg2-binary
celery
redis